refactor: centralize static sign detection
